+++ packages/terra-application/src/breakpoints/index.js
@@ -1,2 +1,4 @@
 export { default } from 'terra-breakpoints';
-export * from 'terra-breakpoints';
+export {
+  ActiveBreakpointContext, ActiveBreakpointProvider, activeBreakpointForSize, breakpointIsActiveForSize, withActiveBreakpoint,
+} from 'terra-breakpoints';